WIP: route_object refactoring
