+++ src/atsc3_logging_externs.h
@@ -171,6 +171,11 @@ extern int _ROUTE_OBJECT_INFO_ENABLED;
 extern int _ROUTE_OBJECT_DEBUG_ENABLED;
 extern int _ROUTE_OBJECT_TRACE_ENABLED;
 
+extern int _SLS_ALC_FLOW_INFO_ENABLED;
+extern int _SLS_ALC_FLOW_DEBUG_ENABLED;
+extern int _SLS_ALC_FLOW_TRACE_ENABLED;
+
+
 //c++ linkage
 //extern int _ISOBMFFTRACKJOINER_DEBUG_ENABLED;
