+++ src/makefile
@@ -414,6 +414,9 @@ atsc3_utils_sha256.o: atsc3_utils_sha256.h atsc3_utils_sha256.c
 atsc3_route_object.o:	atsc3_route_object.h atsc3_route_object.c
 	cc $(CFLAGS) $(DEBUG)  -c atsc3_route_object.c -o atsc3_route_object.o
 	
+atsc3_sls_alc_flow.o:	atsc3_sls_alc_flow.h atsc3_sls_alc_flow.c
+	cc $(CFLAGS) $(DEBUG)  -c atsc3_sls_alc_flow.c -o atsc3_sls_alc_flow.o
+	
 
 # unit standalone tests with mock data
 
@@ -512,7 +515,8 @@ libatsc3_intermediate.o: atsc3_lls.o \
 		atsc3_pcre2_regex_utils.o \
 		atsc3_route_dash_utils.o \
 		atsc3_utils_sha256.o \
-		atsc3_route_object.o
+		atsc3_route_object.o \
+		atsc3_sls_alc_flow.o
 
 	ld  -o libatsc3_intermediate.o -r \
 		atsc3_lls.o \
@@ -573,7 +577,9 @@ libatsc3_intermediate.o: atsc3_lls.o \
 		atsc3_pcre2_regex_utils.o \
 		atsc3_route_dash_utils.o \
 		atsc3_utils_sha256.o \
-		atsc3_route_object.o
+		atsc3_route_object.o \
+		atsc3_sls_alc_flow.o
+		
 
 libatsc3.o: libatsc3_intermediate.o bento4_mock.o
 	ld  -o libatsc3.o -r libatsc3_intermediate.o bento4_mock.o
@@ -889,7 +895,8 @@ test/atsc3_route_dash_mpd_patch_test: test/atsc3_route_dash_mpd_patch_test.c \
 										xml.o \
 										atsc3_mmtp_ntp32_to_pts.o \
 										atsc3_utils_sha256.o \
-										atsc3_route_object.o
+										atsc3_route_object.o \
+										atsc3_sls_alc_flow.o
 										
 										
 	cc $(ASAN) $(CFLAGS) $(DEBUG) test/atsc3_route_dash_mpd_patch_test.c  \
@@ -922,6 +929,7 @@ test/atsc3_route_dash_mpd_patch_test: test/atsc3_route_dash_mpd_patch_test.c \
 							atsc3_mmtp_ntp32_to_pts.o \
 							atsc3_utils_sha256.o \
 							atsc3_route_object.o \
+							atsc3_sls_alc_flow.o \
 							$(LOCAL_LD_LIBS) \
 							-o test/atsc3_route_dash_mpd_patch_test
